feat(core): support for valibot
