+++ packages/core/package.json
@@ -21,10 +21,14 @@
   },
   "peerDependencies": {
     "hono": "^4.6.4",
-    "@hono/zod-validator": "^0.4.1",
-    "zod": "^3.23.8",
-    "zod-openapi": "^4.0.0",
     "openapi-types": "^12.1.3"
+  },
+  "optionalDependencies": {
+    "@hono/valibot-validator": "^0.5.0",
+    "@hono/zod-validator": "^0.4.1",
+    "valibot": "^1.0.0-beta.3",
+    "zod": "^4.0.0",
+    "zod-openapi": "^3.1.1"
   },
   "devDependencies": {
     "@rollup/plugin-terser": "^0.4.4"
