+++ packages/core/rollup.config.js
@@ -1,5 +1,6 @@
 const { withNx } = require("@nx/rollup/with-nx");
 const terser = require("@rollup/plugin-terser");
+const pkg = require("./package.json");
 
 module.exports = withNx(
   {
@@ -9,6 +10,9 @@ module.exports = withNx(
     compiler: "swc",
     format: ["cjs", "esm"],
     assets: [{ input: ".", output: ".", glob: "README.md" }],
+    external: pkg.optionalDependencies
+      ? Object.keys(pkg.optionalDependencies)
+      : [],
   },
   {
     input: {
