--- packages/core/src/constants.ts
@@ -1,1 +1,0 @@
-export const uniqueSymbol = Symbol("openapi");
