+++ packages/core/src/utils.ts
@@ -1,134 +1,49 @@
+import type { ValidationTargets } from "hono";
+import type { ResolverResult } from "./types";
 import type { OpenAPIV3 } from "openapi-types";
-import type { OpenAPIRoute } from "./types";
 
-export const ALLOWED_METHODS = [
-  "GET",
-  "PUT",
-  "POST",
-  "DELETE",
-  "OPTIONS",
-  "HEAD",
-  "PATCH",
-  "TRACE",
-] as const;
+export const uniqueSymbol = Symbol("openapi");
 
-export const toOpenAPIPath = (path: string) =>
-  path
-    .split("/")
-    .map((x) => {
-      let tmp = x;
-      if (tmp.startsWith(":")) {
-        tmp = tmp.slice(1, tmp.length);
-        if (tmp.endsWith("?")) tmp = tmp.slice(0, -1);
-        tmp = `{${tmp}}`;
-      }
-
-      return tmp;
-    })
-    .join("/");
-
-export const capitalize = (word: string) =>
-  word.charAt(0).toUpperCase() + word.slice(1);
-
-export const generateOperationId = (method: string, paths: string) => {
-  let operationId = method;
-
-  if (paths === "/") return `${operationId}Index`;
-
-  for (const path of paths.split("/")) {
-    if (path.charCodeAt(0) === 123) {
-      operationId += `By${capitalize(path.slice(1, -1))}`;
-    } else {
-      operationId += capitalize(path);
-    }
-  }
-
-  return operationId;
-};
-
-export function registerSchemaPath({
-  path,
-  method: _method,
-  data,
-  schema,
-}: OpenAPIRoute & {
-  schema: Partial<OpenAPIV3.PathsObject>;
-}) {
-  path = toOpenAPIPath(path);
-  const method = _method.toLowerCase() as Lowercase<OpenAPIRoute["method"]>;
-
-  schema[path] = {
-    ...(schema[path] ? schema[path] : {}),
-    [method]: {
-      responses: {},
-      ...(schema[path]?.[method] ?? {}),
-      operationId: generateOperationId(method, path),
-      ...data,
-    } satisfies OpenAPIV3.OperationObject,
-  };
-}
-
-export function filterPaths(
-  paths: OpenAPIV3.PathsObject,
-  {
-    excludeStaticFile = true,
-    exclude = [],
-  }: {
-    excludeStaticFile: boolean;
-    exclude: (string | RegExp)[];
-  }
+export function generateValidatorDocs<Target extends keyof ValidationTargets>(
+  target: Target,
+  result: ReturnType<ResolverResult["builder"]>
 ) {
-  const newPaths: OpenAPIV3.PathsObject = {};
+  const docs: Pick<OpenAPIV3.OperationObject, "parameters" | "requestBody"> =
+    {};
 
-  for (const [key, value] of Object.entries(paths)) {
-    if (
-      !exclude.some((x) => {
-        if (typeof x === "string") return key === x;
+  if (target === "form" || target === "json") {
+    docs.requestBody = {
+      content: {
+        [target === "json"
+          ? "application/json"
+          : "application/x-www-form-urlencoded"]: {
+          schema: result.schema,
+        },
+      },
+    };
+  } else {
+    const parameters = [];
 
-        return x.test(key);
-      }) &&
-      !key.includes("*") &&
-      (excludeStaticFile ? !key.includes(".") : true)
-    ) {
-      // @ts-expect-error
-      for (const method of Object.keys(value)) {
-        // @ts-expect-error
-        const schema = value[method];
-
-        if (key.includes("{")) {
-          if (!schema.parameters) schema.parameters = [];
-
-          schema.parameters = [
-            ...key
-              .split("/")
-              .filter(
-                (x) =>
-                  x.startsWith("{") &&
-                  !schema.parameters.find(
-                    (params: Record<string, unknown>) =>
-                      params.in === "path" &&
-                      params.name === x.slice(1, x.length - 1)
-                  )
-              )
-              .map((x) => ({
-                schema: { type: "string" },
-                in: "path",
-                name: x.slice(1, x.length - 1),
-                required: true,
-              })),
-            ...schema.parameters,
-          ];
-        }
-
-        if (!schema.responses)
-          schema.responses = {
-            200: {},
-          };
+    if ("$ref" in result.schema) {
+      parameters.push({
+        in: target,
+        name: result.schema.$ref,
+        schema: result.schema,
+      });
+    } else {
+      for (const [key, value] of Object.entries(
+        result.schema.properties ?? {}
+      )) {
+        parameters.push({
+          in: target,
+          name: key,
+          schema: value,
+        });
       }
-
-      newPaths[key] = value;
     }
+
+    docs.parameters = parameters;
   }
 
-  return newPaths;
+  return { docs, components: result.components };
 }
